print warning on chrome os

+++ install.sh
@@ -6,6 +6,11 @@ if [ "$(id -u)" -eq 0 ]; then
     exit 1
 fi
 
+if grep -q "CHROMEOS_RELEASE_NAME" /etc/lsb-release 2>/dev/null; then
+	echo "ChromeOS is not supported! Use the chrome extension. https://chromewebstore.google.com/detail/vencord-web/cbghhgpcnddeihccjmnadmkaejncjndb"
+	exit 1
+fi
+
 outfile=$(mktemp --tmpdir="$HOME")
 trap 'rm -f "$outfile"' EXIT
